[refactor] QuizCreatePage 단위로 폴더생성 후 관련 컴포넌트 관리 & 스타일 선언 파일 하단으로 이동
